chore: convert domain-redirect integration test to modern integ-runner
