+++ .gitignore
@@ -96,18 +96,5 @@ lib/__tests__/devapp/cdk.out
 !/src/backend/transliterator/transliterator.ts
 !/src/backend/transliterator/transliterator.ecs-entrypoint.ts
 !/lib/backend/transliterator/transliterator.ecs-entrypoint.bundle/Dockerfile
-src/__tests__/domain-redirect.integ.cdkout/asset.*
-src/__tests__/domain-redirect.integ.cdkout/**/asset.*
-src/__tests__/domain-redirect.integ.cdkout/cdk.out
-src/__tests__/domain-redirect.integ.cdkout/**/cdk.out
-src/__tests__/domain-redirect.integ.cdkout/manifest.json
-src/__tests__/domain-redirect.integ.cdkout/**/manifest.json
-src/__tests__/domain-redirect.integ.cdkout/tree.json
-src/__tests__/domain-redirect.integ.cdkout/**/tree.json
-src/__tests__/domain-redirect.integ.cdkout/.cache
-src/__tests__/domain-redirect.integ.cdkout/**/.cache
-src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy
-src/__tests__/.tmp.domain-redirect.integ.cdkout.actual
-!src/__tests__/domain-redirect.integ.cdkout
 !/src/spdx-license.ts
 !/.projenrc.ts
