+++ .npmignore
@@ -27,9 +27,6 @@ tsconfig.tsbuildinfo
 /test
 !/website
 /update-github-ip-allowlist.js
-src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy
-src/__tests__/.tmp.domain-redirect.integ.cdkout.actual
-src/__tests__/domain-redirect.integ.cdkout
 /.gitattributes
 /.projenrc.ts
 /projenrc
