+++ .projen/tasks.json
@@ -911,57 +911,6 @@
         }
       ]
     },
-    "integ:domain-redirect:assert": {
-      "name": "integ:domain-redirect:assert",
-      "description": "synthesize integration test __tests__/domain-redirect.integ.ts",
-      "steps": [
-        {
-          "exec": "cdk synth --app \"node lib/__tests__/domain-redirect.integ.js\" --no-version-reporting --context @aws-cdk/core:newStyleStackSynthesis=true -o src/__tests__/.tmp.domain-redirect.integ.cdkout.actual > /dev/null"
-        },
-        {
-          "exec": "diff -r -x asset.* -x cdk.out -x manifest.json -x tree.json -x .cache src/__tests__/domain-redirect.integ.cdkout/ src/__tests__/.tmp.domain-redirect.integ.cdkout.actual/"
-        }
-      ]
-    },
-    "integ:domain-redirect:deploy": {
-      "name": "integ:domain-redirect:deploy",
-      "description": "deploy integration test __tests__/domain-redirect.integ.ts",
-      "steps": [
-        {
-          "exec": "rm -fr src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy"
-        },
-        {
-          "exec": "cdk deploy --app \"node lib/__tests__/domain-redirect.integ.js\" --no-version-reporting --context @aws-cdk/core:newStyleStackSynthesis=true --require-approval=never -o src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy"
-        },
-        {
-          "exec": "rm -fr src/__tests__/domain-redirect.integ.cdkout"
-        },
-        {
-          "exec": "mv src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy src/__tests__/domain-redirect.integ.cdkout"
-        },
-        {
-          "spawn": "integ:domain-redirect:destroy"
-        }
-      ]
-    },
-    "integ:domain-redirect:destroy": {
-      "name": "integ:domain-redirect:destroy",
-      "description": "destroy integration test __tests__/domain-redirect.integ.ts",
-      "steps": [
-        {
-          "exec": "cdk destroy --app src/__tests__/domain-redirect.integ.cdkout --no-version-reporting"
-        }
-      ]
-    },
-    "integ:domain-redirect:snapshot": {
-      "name": "integ:domain-redirect:snapshot",
-      "description": "update snapshot for integration test __tests__/domain-redirect.integ.ts",
-      "steps": [
-        {
-          "exec": "cdk synth --app \"node lib/__tests__/domain-redirect.integ.js\" --no-version-reporting --context @aws-cdk/core:newStyleStackSynthesis=true -o src/__tests__/domain-redirect.integ.cdkout > /dev/null"
-        }
-      ]
-    },
     "integ:update": {
       "name": "integ:update",
       "description": "Run and update integration snapshot tests",
@@ -1070,9 +1019,6 @@
         },
         {
           "spawn": "integ"
-        },
-        {
-          "spawn": "integ:domain-redirect:assert"
         }
       ]
     },
