+++ package.json
@@ -79,10 +79,6 @@
     "eject": "npx projen eject",
     "eslint": "npx projen eslint",
     "integ": "npx projen integ",
-    "integ:domain-redirect:assert": "npx projen integ:domain-redirect:assert",
-    "integ:domain-redirect:deploy": "npx projen integ:domain-redirect:deploy",
-    "integ:domain-redirect:destroy": "npx projen integ:domain-redirect:destroy",
-    "integ:domain-redirect:snapshot": "npx projen integ:domain-redirect:snapshot",
     "integ:update": "npx projen integ:update",
     "package": "npx projen package",
     "package-all": "npx projen package-all",
