+++ test/integ.domain-redirect.ts.snapshot/integ.json
@@ -1,0 +1,12 @@
+{
+  "version": "36.0.0",
+  "testCases": {
+    "domain-redirect-integ/DefaultTest": {
+      "stacks": [
+        "DomainRedirectIntegrationTest"
+      ],
+      "assertionStack": "domain-redirect-integ/DefaultTest/DeployAssert",
+      "assertionStackName": "domainredirectintegDefaultTestDeployAssertD715F788"
+    }
+  }
+}
